requirements.txt
astroid==2.3.2
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
configobj==5.0.6
decorator==4.0.6
defusedxml==0.6.0
Django==2.2.6
django-appconf==1.0.3
django-autoslug-iplweb==1.9.4.dev0
django-axes==5.0.20
django-cities-light==3.5.0
django-countries==5.5
django-datatable-view==0.9.0
django-datatables-view==1.19.1
django-formtools==2.1
django-ipware==2.1.0
django-map-widgets==0.2.2
django-otp==0.7.3
django-phonenumber-field==1.3.0
django-two-factor-auth==1.9.1
django-user-activity-log2==0.0.24
dulwich==0.12.0
fastimport==0.9.4
idna==2.8
ipython==2.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.5
oauthlib==3.1.0
pexpect==4.0.1
phonenumberslite==8.10.23
Pillow==6.2.1
progressbar2==3.47.0
ptyprocess==0.5
pycodestyle==2.5.0
pycryptodome==3.9.4
PyJWT==1.7.1
pylint==2.4.3
python-dateutil==2.8.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.3
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.2.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
twilio==6.33.1
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
wrapt==1.11.2
